Revert "Adds Hilt to the project"
