--- app/src/androidTest/java/com/example/android/architecture/blueprints/todoapp/AppNavigationTest.kt
@@ -1,246 +1,0 @@
-/*
- * Copyright (C) 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.architecture.blueprints.todoapp
-
-import android.view.Gravity
-import androidx.drawerlayout.widget.DrawerLayout
-import androidx.navigation.findNavController
-import androidx.test.core.app.ActivityScenario
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.Espresso.pressBack
-import androidx.test.espresso.IdlingRegistry
-import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.contrib.DrawerActions.open
-import androidx.test.espresso.contrib.DrawerMatchers.isClosed
-import androidx.test.espresso.contrib.DrawerMatchers.isOpen
-import androidx.test.espresso.contrib.NavigationViewActions.navigateTo
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withContentDescription
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import androidx.test.espresso.matcher.ViewMatchers.withText
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.filters.LargeTest
-import com.example.android.architecture.blueprints.todoapp.data.Task
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
-import com.example.android.architecture.blueprints.todoapp.di.TasksRepositoryModule
-import com.example.android.architecture.blueprints.todoapp.tasks.TasksActivity
-import com.example.android.architecture.blueprints.todoapp.util.EspressoIdlingResource
-import com.example.android.architecture.blueprints.todoapp.util.saveTaskBlocking
-import dagger.hilt.android.testing.HiltAndroidRule
-import dagger.hilt.android.testing.HiltAndroidTest
-import dagger.hilt.android.testing.UninstallModules
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import javax.inject.Inject
-
-/**
- * Tests for the [DrawerLayout] layout component in [TasksActivity] which manages
- * navigation within the app.
- *
- * UI tests usually use [ActivityTestRule] but there's no API to perform an action before
- * each test. The workaround is to use `ActivityScenario.launch()` and `ActivityScenario.close()`.
- */
-@RunWith(AndroidJUnit4::class)
-@LargeTest
-@UninstallModules(TasksRepositoryModule::class)
-@HiltAndroidTest
-class AppNavigationTest {
-
-    @get:Rule
-    var hiltRule = HiltAndroidRule(this)
-
-    @Inject
-    lateinit var tasksRepository: TasksRepository
-
-    // An Idling Resource that waits for Data Binding to have no pending bindings
-    private val dataBindingIdlingResource = DataBindingIdlingResource()
-
-    @Before
-    fun init() {
-        // Populate @Inject fields in test class
-        hiltRule.inject()
-    }
-
-    /**
-     * Idling resources tell Espresso that the app is idle or busy. This is needed when operations
-     * are not scheduled in the main Looper (for example when executed on a different thread).
-     */
-    @Before
-    fun registerIdlingResource() {
-        IdlingRegistry.getInstance().register(EspressoIdlingResource.countingIdlingResource)
-        IdlingRegistry.getInstance().register(dataBindingIdlingResource)
-    }
-
-    /**
-     * Unregister your Idling Resource so it can be garbage collected and does not leak any memory.
-     */
-    @After
-    fun unregisterIdlingResource() {
-        IdlingRegistry.getInstance().unregister(EspressoIdlingResource.countingIdlingResource)
-        IdlingRegistry.getInstance().unregister(dataBindingIdlingResource)
-    }
-
-    @Test
-    fun drawerNavigationFromTasksToStatistics() {
-        // start up Tasks screen
-        val activityScenario = ActivityScenario.launch(TasksActivity::class.java)
-        dataBindingIdlingResource.monitorActivity(activityScenario)
-
-        onView(withId(R.id.drawer_layout))
-            .check(matches(isClosed(Gravity.START))) // Left Drawer should be closed.
-            .perform(open()) // Open Drawer
-
-        // Start statistics screen.
-        onView(withId(R.id.nav_view))
-            .perform(navigateTo(R.id.statistics_fragment_dest))
-
-        // Check that statistics screen was opened.
-        onView(withId(R.id.statistics_layout)).check(matches(isDisplayed()))
-
-        onView(withId(R.id.drawer_layout))
-            .check(matches(isClosed(Gravity.START))) // Left Drawer should be closed.
-            .perform(open()) // Open Drawer
-
-        // Start tasks screen.
-        onView(withId(R.id.nav_view))
-            .perform(navigateTo(R.id.tasks_fragment_dest))
-
-        // Check that tasks screen was opened.
-        onView(withId(R.id.tasks_container_layout)).check(matches(isDisplayed()))
-        // When using ActivityScenario.launch, always call close()
-        activityScenario.close()
-    }
-
-    @Test
-    fun tasksScreen_clickOnAndroidHomeIcon_OpensNavigation() {
-        // start up Tasks screen
-        val activityScenario = ActivityScenario.launch(TasksActivity::class.java)
-        dataBindingIdlingResource.monitorActivity(activityScenario)
-
-        // Check that left drawer is closed at startup
-        onView(withId(R.id.drawer_layout))
-            .check(matches(isClosed(Gravity.START))) // Left Drawer should be closed.
-
-        // Open Drawer
-        onView(
-            withContentDescription(
-                activityScenario
-                    .getToolbarNavigationContentDescription()
-            )
-        ).perform(click())
-
-        // Check if drawer is open
-        onView(withId(R.id.drawer_layout))
-            .check(matches(isOpen(Gravity.START))) // Left drawer is open open.
-        // When using ActivityScenario.launch, always call close()
-        activityScenario.close()
-    }
-
-    @Test
-    fun statsScreen_clickOnAndroidHomeIcon_OpensNavigation() {
-        // start up Tasks screen
-        val activityScenario = ActivityScenario.launch(TasksActivity::class.java)
-        dataBindingIdlingResource.monitorActivity(activityScenario)
-
-        // When the user navigates to the stats screen
-        activityScenario.onActivity {
-            it.findNavController(R.id.nav_host_fragment).navigate(R.id.statistics_fragment_dest)
-        }
-
-        // Then check that left drawer is closed at startup
-        onView(withId(R.id.drawer_layout))
-            .check(matches(isClosed(Gravity.START))) // Left Drawer should be closed.
-
-        // When the drawer is opened
-        onView(
-            withContentDescription(
-                activityScenario
-                    .getToolbarNavigationContentDescription()
-            )
-        ).perform(click())
-
-        // Then check that the drawer is open
-        onView(withId(R.id.drawer_layout))
-            .check(matches(isOpen(Gravity.START))) // Left drawer is open open.
-        // When using ActivityScenario.launch, always call close()
-        activityScenario.close()
-    }
-
-    @Test
-    fun taskDetailScreen_doubleUIBackButton() {
-        val task = Task("UI <- button", "Description")
-        tasksRepository.saveTaskBlocking(task)
-
-        // start up Tasks screen
-        val activityScenario = ActivityScenario.launch(TasksActivity::class.java)
-        dataBindingIdlingResource.monitorActivity(activityScenario)
-
-        // Click on the task on the list
-        onView(withText("UI <- button")).perform(click())
-        // Click on the edit task button
-        onView(withId(R.id.edit_task_fab)).perform(click())
-
-        // Confirm that if we click "<-" once, we end up back at the task details page
-        onView(
-            withContentDescription(
-                activityScenario
-                    .getToolbarNavigationContentDescription()
-            )
-        ).perform(click())
-        onView(withId(R.id.task_detail_title_text)).check(matches(isDisplayed()))
-
-        // Confirm that if we click "<-" a second time, we end up back at the home screen
-        onView(
-            withContentDescription(
-                activityScenario
-                    .getToolbarNavigationContentDescription()
-            )
-        ).perform(click())
-        onView(withId(R.id.tasks_container_layout)).check(matches(isDisplayed()))
-        // When using ActivityScenario.launch, always call close()
-        activityScenario.close()
-    }
-
-    @Test
-    fun taskDetailScreen_doubleBackButton() {
-        val task = Task("Back button", "Description")
-        tasksRepository.saveTaskBlocking(task)
-
-        // start up Tasks screen
-        val activityScenario = ActivityScenario.launch(TasksActivity::class.java)
-        dataBindingIdlingResource.monitorActivity(activityScenario)
-
-        // Click on the task on the list
-        onView(withText("Back button")).perform(click())
-        // Click on the edit task button
-        onView(withId(R.id.edit_task_fab)).perform(click())
-
-        // Confirm that if we click back once, we end up back at the task details page
-        pressBack()
-        onView(withId(R.id.task_detail_title_text)).check(matches(isDisplayed()))
-
-        // Confirm that if we click back a second time, we end up back at the home screen
-        pressBack()
-        onView(withId(R.id.tasks_container_layout)).check(matches(isDisplayed()))
-        // When using ActivityScenario.launch, always call close()
-        activityScenario.close()
-    }
-}
