--- app/src/androidTest/java/com/example/android/architecture/blueprints/todoapp/DataBindingIdlingResource.kt
@@ -1,112 +1,0 @@
-/*
- * Copyright (C) 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.architecture.blueprints.todoapp
-
-import android.view.View
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
-import androidx.fragment.app.testing.FragmentScenario
-import androidx.test.core.app.ActivityScenario
-import androidx.test.espresso.IdlingResource
-import java.util.UUID
-
-/**
- * An espresso idling resource implementation that reports idle status for all data binding
- * layouts. Data Binding uses a mechanism to post messages which Espresso doesn't track yet.
- *
- * Since this application only uses fragments, the resource only checks the fragments and their
- * children instead of the whole view tree.
- */
-class DataBindingIdlingResource : IdlingResource {
-    // list of registered callbacks
-    private val idlingCallbacks = mutableListOf<IdlingResource.ResourceCallback>()
-    // give it a unique id to workaround an espresso bug where you cannot register/unregister
-    // an idling resource w/ the same name.
-    private val id = UUID.randomUUID().toString()
-    // holds whether isIdle is called and the result was false. We track this to avoid calling
-    // onTransitionToIdle callbacks if Espresso never thought we were idle in the first place.
-    private var wasNotIdle = false
-
-    lateinit var activity: FragmentActivity
-
-    override fun getName() = "DataBinding $id"
-
-    override fun isIdleNow(): Boolean {
-        val idle = !getBindings().any { it.hasPendingBindings() }
-        @Suppress("LiftReturnOrAssignment")
-        if (idle) {
-            if (wasNotIdle) {
-                // notify observers to avoid espresso race detector
-                idlingCallbacks.forEach { it.onTransitionToIdle() }
-            }
-            wasNotIdle = false
-        } else {
-            wasNotIdle = true
-            // check next frame
-            activity.findViewById<View>(android.R.id.content).postDelayed({
-                isIdleNow
-            }, 16)
-        }
-        return idle
-    }
-
-    override fun registerIdleTransitionCallback(callback: IdlingResource.ResourceCallback) {
-        idlingCallbacks.add(callback)
-    }
-
-    /**
-     * Find all binding classes in all currently available fragments.
-     */
-    private fun getBindings(): List<ViewDataBinding> {
-        val fragments = (activity as? FragmentActivity)
-            ?.supportFragmentManager
-            ?.fragments
-
-        val bindings =
-            fragments?.mapNotNull {
-                it.view?.getBinding()
-            } ?: emptyList()
-        val childrenBindings = fragments?.flatMap { it.childFragmentManager.fragments }
-            ?.mapNotNull { it.view?.getBinding() } ?: emptyList()
-
-        return bindings + childrenBindings
-    }
-}
-
-private fun View.getBinding(): ViewDataBinding? = DataBindingUtil.getBinding(this)
-
-/**
- * Sets the activity from an [ActivityScenario] to be used from [DataBindingIdlingResource].
- */
-fun DataBindingIdlingResource.monitorActivity(
-    activityScenario: ActivityScenario<out FragmentActivity>
-) {
-    activityScenario.onActivity {
-        this.activity = it
-    }
-}
-
-/**
- * Sets the fragment from a [FragmentScenario] to be used from [DataBindingIdlingResource].
- */
-fun DataBindingIdlingResource.monitorFragment(fragmentScenario: FragmentScenario<out Fragment>) {
-    fragmentScenario.onFragment {
-        this.activity = it.requireActivity()
-    }
-}
