--- app/src/androidTest/java/com/example/android/architecture/blueprints/todoapp/taskdetail/TaskDetailFragmentTest.kt
@@ -1,106 +1,0 @@
-/*
- * Copyright (C) 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.example.android.architecture.blueprints.todoapp.taskdetail
-
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers.isChecked
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import androidx.test.espresso.matcher.ViewMatchers.withText
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.filters.MediumTest
-import com.example.android.architecture.blueprints.todoapp.R
-import com.example.android.architecture.blueprints.todoapp.data.Task
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
-import com.example.android.architecture.blueprints.todoapp.di.TasksRepositoryModule
-import com.example.android.architecture.blueprints.todoapp.launchFragmentInHiltContainer
-import com.example.android.architecture.blueprints.todoapp.util.saveTaskBlocking
-import dagger.hilt.android.testing.HiltAndroidRule
-import dagger.hilt.android.testing.HiltAndroidTest
-import dagger.hilt.android.testing.UninstallModules
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import org.hamcrest.core.IsNot.not
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import javax.inject.Inject
-
-/**
- * Integration test for the Task Details screen.
- */
-@MediumTest
-@RunWith(AndroidJUnit4::class)
-@ExperimentalCoroutinesApi
-@UninstallModules(TasksRepositoryModule::class)
-@HiltAndroidTest
-class TaskDetailFragmentTest {
-
-    @get:Rule
-    var hiltRule = HiltAndroidRule(this)
-
-    @Inject
-    lateinit var repository: TasksRepository
-
-    @Before
-    fun init() {
-        // Populate @Inject fields in test class
-        hiltRule.inject()
-    }
-
-    @Test
-    fun activeTaskDetails_DisplayedInUi() {
-        // GIVEN - Add active (incomplete) task to the DB
-        val activeTask = Task("Active Task", "AndroidX Rocks", false)
-        repository.saveTaskBlocking(activeTask)
-
-        // WHEN - Details fragment launched to display task
-        val bundle = TaskDetailFragmentArgs(activeTask.id).toBundle()
-        launchFragmentInHiltContainer<TaskDetailFragment>(bundle, R.style.AppTheme)
-
-        // THEN - Task details are displayed on the screen
-        // make sure that the title/description are both shown and correct
-        onView(withId(R.id.task_detail_title_text)).check(matches(isDisplayed()))
-        onView(withId(R.id.task_detail_title_text)).check(matches(withText("Active Task")))
-        onView(withId(R.id.task_detail_description_text)).check(matches(isDisplayed()))
-        onView(withId(R.id.task_detail_description_text)).check(matches(withText("AndroidX Rocks")))
-        // and make sure the "active" checkbox is shown unchecked
-        onView(withId(R.id.task_detail_complete_checkbox)).check(matches(isDisplayed()))
-        onView(withId(R.id.task_detail_complete_checkbox)).check(matches(not(isChecked())))
-    }
-
-    @Test
-    fun completedTaskDetails_DisplayedInUi() {
-        // GIVEN - Add completed task to the DB
-        val completedTask = Task("Completed Task", "AndroidX Rocks", true)
-        repository.saveTaskBlocking(completedTask)
-
-        // WHEN - Details fragment launched to display task
-        val bundle = TaskDetailFragmentArgs(completedTask.id).toBundle()
-        launchFragmentInHiltContainer<TaskDetailFragment>(bundle, R.style.AppTheme)
-
-        // THEN - Task details are displayed on the screen
-        // make sure that the title/description are both shown and correct
-        onView(withId(R.id.task_detail_title_text)).check(matches(isDisplayed()))
-        onView(withId(R.id.task_detail_title_text)).check(matches(withText("Completed Task")))
-        onView(withId(R.id.task_detail_description_text)).check(matches(isDisplayed()))
-        onView(withId(R.id.task_detail_description_text)).check(matches(withText("AndroidX Rocks")))
-        // and make sure the "active" checkbox is shown unchecked
-        onView(withId(R.id.task_detail_complete_checkbox)).check(matches(isDisplayed()))
-        onView(withId(R.id.task_detail_complete_checkbox)).check(matches(isChecked()))
-    }
-}
