+++ app/src/androidTestMock/java/com/example/android/architecture/blueprints/todoapp/tasks/TestUtils.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright (C) 2019 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.example.android.architecture.blueprints.todoapp.tasks
+
+import android.app.Activity
+import androidx.appcompat.widget.Toolbar
+import androidx.test.core.app.ActivityScenario
+import com.example.android.architecture.blueprints.todoapp.R
+
+fun <T : Activity> ActivityScenario<T>.getToolbarNavigationContentDescription(): String {
+    var description = ""
+    onActivity {
+        description =
+            it.findViewById<Toolbar>(R.id.toolbar).navigationContentDescription as String
+    }
+    return description
+}
