+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/TodoApplication.kt
@@ -17,18 +17,21 @@
 package com.example.android.architecture.blueprints.todoapp
 
 import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
+import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
 import timber.log.Timber
 import timber.log.Timber.DebugTree
 
 /**
- * An application with @HiltAndroidApp that triggers Hilt's code generation and
- * adds an application-level dependency container.
+ * An application that lazily provides a repository. Note that this Service Locator pattern is
+ * used to simplify the sample. Consider a Dependency Injection framework.
  *
  * Also, sets up Timber in the DEBUG BuildConfig. Read Timber's documentation for production setups.
  */
-@HiltAndroidApp
 class TodoApplication : Application() {
+
+    // Depends on the flavor,
+    val taskRepository: TasksRepository
+        get() = ServiceLocator.provideTasksRepository(this)
 
     override fun onCreate() {
         super.onCreate()
