+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/addedittask/AddEditTaskViewModel.kt
@@ -16,7 +16,6 @@
 
 package com.example.android.architecture.blueprints.todoapp.addedittask
 
-import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -31,7 +30,7 @@ import kotlinx.coroutines.launch
 /**
  * ViewModel for the Add/Edit screen.
  */
-class AddEditTaskViewModel @ViewModelInject constructor(
+class AddEditTaskViewModel(
     private val tasksRepository: TasksRepository
 ) : ViewModel() {
