--- app/src/main/java/com/example/android/architecture/blueprints/todoapp/di/AppModule.kt
@@ -1,108 +1,0 @@
-/*
- * Copyright (C) 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.architecture.blueprints.todoapp.di
-
-import android.content.Context
-import androidx.room.Room
-import com.example.android.architecture.blueprints.todoapp.data.source.DefaultTasksRepository
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
-import com.example.android.architecture.blueprints.todoapp.data.source.local.TasksLocalDataSource
-import com.example.android.architecture.blueprints.todoapp.data.source.local.ToDoDatabase
-import com.example.android.architecture.blueprints.todoapp.data.source.remote.TasksRemoteDataSource
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ApplicationComponent
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import javax.inject.Qualifier
-import javax.inject.Singleton
-import kotlin.annotation.AnnotationRetention.RUNTIME
-
-/**
- * Module to tell Hilt how to provide instances of types that cannot be constructor-injected.
- *
- * As these types are scoped to the application lifecycle using @Singleton, they're installed
- * in Hilt's ApplicationComponent.
- */
-@Module
-@InstallIn(ApplicationComponent::class)
-object AppModule {
-
-    @Qualifier
-    @Retention(RUNTIME)
-    annotation class RemoteTasksDataSource
-
-    @Qualifier
-    @Retention(RUNTIME)
-    annotation class LocalTasksDataSource
-
-    @Singleton
-    @RemoteTasksDataSource
-    @Provides
-    fun provideTasksRemoteDataSource(): TasksDataSource {
-        return TasksRemoteDataSource
-    }
-
-    @Singleton
-    @LocalTasksDataSource
-    @Provides
-    fun provideTasksLocalDataSource(
-        database: ToDoDatabase,
-        ioDispatcher: CoroutineDispatcher
-    ): TasksDataSource {
-        return TasksLocalDataSource(
-            database.taskDao(), ioDispatcher
-        )
-    }
-
-    @Singleton
-    @Provides
-    fun provideDataBase(@ApplicationContext context: Context): ToDoDatabase {
-        return Room.databaseBuilder(
-            context.applicationContext,
-            ToDoDatabase::class.java,
-            "Tasks.db"
-        ).build()
-    }
-
-    @Singleton
-    @Provides
-    fun provideIoDispatcher() = Dispatchers.IO
-}
-
-/**
- * The binding for TasksRepository is on its own module so that we can replace it easily in tests.
- */
-@Module
-@InstallIn(ApplicationComponent::class)
-object TasksRepositoryModule {
-
-    @Singleton
-    @Provides
-    fun provideTasksRepository(
-        @AppModule.RemoteTasksDataSource remoteTasksDataSource: TasksDataSource,
-        @AppModule.LocalTasksDataSource localTasksDataSource: TasksDataSource,
-        ioDispatcher: CoroutineDispatcher
-    ): TasksRepository {
-        return DefaultTasksRepository(
-            remoteTasksDataSource, localTasksDataSource, ioDispatcher
-        )
-    }
-}
