+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/statistics/StatisticsFragment.kt
@@ -24,18 +24,17 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.example.android.architecture.blueprints.todoapp.R
 import com.example.android.architecture.blueprints.todoapp.databinding.StatisticsFragBinding
+import com.example.android.architecture.blueprints.todoapp.util.getViewModelFactory
 import com.example.android.architecture.blueprints.todoapp.util.setupRefreshLayout
-import dagger.hilt.android.AndroidEntryPoint
 
 /**
  * Main UI for the statistics screen.
  */
-@AndroidEntryPoint
 class StatisticsFragment : Fragment() {
 
     private lateinit var viewDataBinding: StatisticsFragBinding
 
-    private val viewModel by viewModels<StatisticsViewModel>()
+    private val viewModel by viewModels<StatisticsViewModel> { getViewModelFactory() }
 
     override fun onCreateView(
         inflater: LayoutInflater,
