+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/statistics/StatisticsViewModel.kt
@@ -16,7 +16,6 @@
 
 package com.example.android.architecture.blueprints.todoapp.statistics
 
-import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -32,7 +31,7 @@ import kotlinx.coroutines.launch
 /**
  * ViewModel for the statistics screen.
  */
-class StatisticsViewModel @ViewModelInject constructor(
+class StatisticsViewModel(
     private val tasksRepository: TasksRepository
 ) : ViewModel() {
