+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/TasksActivity.kt
@@ -27,12 +27,10 @@ import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.example.android.architecture.blueprints.todoapp.R
 import com.google.android.material.navigation.NavigationView
-import dagger.hilt.android.AndroidEntryPoint
 
 /**
  * Main activity for the todoapp. Holds the Navigation Host Fragment and the Drawer, Toolbar, etc.
  */
-@AndroidEntryPoint
 class TasksActivity : AppCompatActivity() {
 
     private lateinit var drawerLayout: DrawerLayout
