+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/TasksFragment.kt
@@ -32,20 +32,19 @@ import com.example.android.architecture.blueprints.todoapp.EventObserver
 import com.example.android.architecture.blueprints.todoapp.R
 import com.example.android.architecture.blueprints.todoapp.data.Task
 import com.example.android.architecture.blueprints.todoapp.databinding.TasksFragBinding
+import com.example.android.architecture.blueprints.todoapp.util.getViewModelFactory
 import com.example.android.architecture.blueprints.todoapp.util.setupRefreshLayout
 import com.example.android.architecture.blueprints.todoapp.util.setupSnackbar
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.snackbar.Snackbar
-import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
 
 /**
  * Display a grid of [Task]s. User can choose to view all, active or completed tasks.
  */
-@AndroidEntryPoint
 class TasksFragment : Fragment() {
 
-    private val viewModel by viewModels<TasksViewModel>()
+    private val viewModel by viewModels<TasksViewModel> { getViewModelFactory() }
 
     private val args: TasksFragmentArgs by navArgs()
 
@@ -99,10 +98,10 @@ class TasksFragment : Fragment() {
     }
 
     private fun setupNavigation() {
-        viewModel.openTaskEvent.observe(viewLifecycleOwner, EventObserver {
+        viewModel.openTaskEvent.observe(this, EventObserver {
             openTaskDetails(it)
         })
-        viewModel.newTaskEvent.observe(viewLifecycleOwner, EventObserver {
+        viewModel.newTaskEvent.observe(this, EventObserver {
             navigateToAddNewTask()
         })
     }
