+++ app/src/mock/java/com/example/android/architecture/blueprints/todoapp/data/FakeTasksRemoteDataSource.kt
@@ -13,36 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.android.architecture.blueprints.todoapp.data.source
+package com.example.android.architecture.blueprints.todoapp.data
 
-import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.map
-import com.example.android.architecture.blueprints.todoapp.data.Result
 import com.example.android.architecture.blueprints.todoapp.data.Result.Error
 import com.example.android.architecture.blueprints.todoapp.data.Result.Success
-import com.example.android.architecture.blueprints.todoapp.data.Task
-import kotlinx.coroutines.runBlocking
+import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource
 import java.util.LinkedHashMap
 
 /**
  * Implementation of a remote data source with static access to the data for easy testing.
  */
-class FakeRepository : TasksRepository {
+object FakeTasksRemoteDataSource : TasksDataSource {
 
-    var tasksServiceData: LinkedHashMap<String, Task> = LinkedHashMap()
-
-    private var shouldReturnError = false
+    private var TASKS_SERVICE_DATA: LinkedHashMap<String, Task> = LinkedHashMap()
 
     private val observableTasks = MutableLiveData<Result<List<Task>>>()
 
-    fun setReturnError(value: Boolean) {
-        shouldReturnError = value
-    }
-
     override suspend fun refreshTasks() {
-        observableTasks.value = getTasks()
+        observableTasks.postValue(getTasks())
     }
 
     override suspend fun refreshTask(taskId: String) {
@@ -50,12 +41,10 @@ class FakeRepository : TasksRepository {
     }
 
     override fun observeTasks(): LiveData<Result<List<Task>>> {
-        runBlocking { refreshTasks() }
         return observableTasks
     }
 
     override fun observeTask(taskId: String): LiveData<Result<Task>> {
-        runBlocking { refreshTasks() }
         return observableTasks.map { tasks ->
             when (tasks) {
                 is Result.Loading -> Result.Loading
@@ -69,67 +58,52 @@ class FakeRepository : TasksRepository {
         }
     }
 
-    override suspend fun getTask(taskId: String, forceUpdate: Boolean): Result<Task> {
-        if (shouldReturnError) {
-            return Error(Exception("Test exception"))
-        }
-        tasksServiceData[taskId]?.let {
+    override suspend fun getTask(taskId: String): Result<Task> {
+        TASKS_SERVICE_DATA[taskId]?.let {
             return Success(it)
         }
         return Error(Exception("Could not find task"))
     }
 
-    override suspend fun getTasks(forceUpdate: Boolean): Result<List<Task>> {
-        if (shouldReturnError) {
-            return Error(Exception("Test exception"))
-        }
-        return Success(tasksServiceData.values.toList())
+    override suspend fun getTasks(): Result<List<Task>> {
+        return Success(TASKS_SERVICE_DATA.values.toList())
     }
 
     override suspend fun saveTask(task: Task) {
-        tasksServiceData[task.id] = task
+        TASKS_SERVICE_DATA[task.id] = task
     }
 
     override suspend fun completeTask(task: Task) {
         val completedTask = Task(task.title, task.description, true, task.id)
-        tasksServiceData[task.id] = completedTask
+        TASKS_SERVICE_DATA[task.id] = completedTask
     }
 
     override suspend fun completeTask(taskId: String) {
         // Not required for the remote data source.
-        throw NotImplementedError()
     }
 
     override suspend fun activateTask(task: Task) {
         val activeTask = Task(task.title, task.description, false, task.id)
-        tasksServiceData[task.id] = activeTask
+        TASKS_SERVICE_DATA[task.id] = activeTask
     }
 
     override suspend fun activateTask(taskId: String) {
-        throw NotImplementedError()
+        // Not required for the remote data source.
     }
 
     override suspend fun clearCompletedTasks() {
-        tasksServiceData = tasksServiceData.filterValues {
+        TASKS_SERVICE_DATA = TASKS_SERVICE_DATA.filterValues {
             !it.isCompleted
         } as LinkedHashMap<String, Task>
     }
 
     override suspend fun deleteTask(taskId: String) {
-        tasksServiceData.remove(taskId)
+        TASKS_SERVICE_DATA.remove(taskId)
         refreshTasks()
     }
 
     override suspend fun deleteAllTasks() {
-        tasksServiceData.clear()
+        TASKS_SERVICE_DATA.clear()
         refreshTasks()
-    }
-
-    @VisibleForTesting
-    fun addTasks(vararg tasks: Task) {
-        for (task in tasks) {
-            tasksServiceData[task.id] = task
-        }
-        runBlocking { refreshTasks() }
     }
 }
