--- app/src/sharedTest/java/com/example/android/architecture/blueprints/todoapp/di/TestTasksRepositoryModule.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright (C) 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.architecture.blueprints.todoapp.di
-
-import com.example.android.architecture.blueprints.todoapp.data.source.DefaultTasksRepository
-import com.example.android.architecture.blueprints.todoapp.data.source.FakeRepository
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ApplicationComponent
-import javax.inject.Singleton
-
-/**
- * TasksRepository binding to use in tests.
- *
- * Hilt will inject a [FakeRepository] instead of a [DefaultTasksRepository].
- */
-@Module
-@InstallIn(ApplicationComponent::class)
-abstract class TestTasksRepositoryModule {
-    @Singleton
-    @Binds
-    abstract fun bindRepository(repo: FakeRepository): TasksRepository
-}
