+++ app/src/test/java/com/example/android/architecture/blueprints/todoapp/LiveDataTestUtil.kt
@@ -53,6 +53,7 @@ fun <T> LiveData<T>.getOrAwaitValue(
             this.removeObserver(observer)
             throw TimeoutException("LiveData value was never set.")
         }
+
     } finally {
         this.removeObserver(observer)
     }
