+++ gradle.properties
@@ -18,4 +18,6 @@
 # org.gradle.parallel=true
 android.enableJetifier=true
 android.useAndroidX=true
+android.enableUnitTestBinaryResources=true
+android.enableR8=true
 kapt.incremental.apt=true
